Fixed Index

+++ src/main/java/com/ponshankar/hackathon/blackrock/model/Expense.java
@@ -4,5 +4,5 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record Expense(
         @JsonProperty("timestamp") String timestamp,
-        @JsonProperty("amount") Long amount
+        @JsonProperty("amount") Double amount
 ) {}
