+++ src/main/java/com/ponshankar/hackathon/blackrock/model/Totals.java
@@ -3,7 +3,7 @@ package com.ponshankar.hackathon.blackrock.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record Totals(
-        @JsonProperty("amountSum") Long amountSum,
-        @JsonProperty("ceilingSum") Long ceilingSum,
-        @JsonProperty("remanentSum") Long remanentSum
+        @JsonProperty("amountSum") Double amountSum,
+        @JsonProperty("ceilingSum") Double ceilingSum,
+        @JsonProperty("remanentSum") Double remanentSum
 ) {}
