+++ src/main/java/com/ponshankar/hackathon/blackrock/model/Transaction.java
@@ -6,12 +6,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record Transaction(
         @JsonProperty("date") String date,
-        @JsonProperty("amount") Long amount,
-        @JsonProperty("ceiling") Long ceiling,
-        @JsonProperty("remanent") Long remanent,
+        @JsonProperty("amount") Double amount,
+        @JsonProperty("ceiling") Double ceiling,
+        @JsonProperty("remanent") Double remanent,
         @JsonProperty("message") String message
 ) {
-    public Transaction(String date, Long amount, Long ceiling, Long remanent) {
+    public Transaction(String date, Double amount, Double ceiling, Double remanent) {
         this(date, amount, ceiling, remanent, null);
     }
 }
