+++ src/main/java/com/ponshankar/hackathon/blackrock/model/request/ValidatorRequest.java
@@ -6,6 +6,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
 public record ValidatorRequest(
-        @JsonProperty("wage") Long wage,
+        @JsonProperty("wage") Double wage,
         @JsonProperty("transactions") List<Transaction> transactions
 ) {}
